rename mockSvmCpiEventsClient file
