+++ src/clients/mocks/index.ts
@@ -1,5 +1,5 @@
 export * from "./MockConfigStoreClient";
 export * from "./MockHubPoolClient";
 export * from "./MockSpokePoolClient";
-export * from "./MockSolanaEventClient";
+export * from "./MockSvmCpiEventsClient";
 export * from "./MockSvmSpokePoolClient";
